==> Source/DataSetCompeticao/1-Braga_Transformation.ipynb <==
<center>
<h1 style="font-family: 'Times New Roman', Times, serif; font-size: 60px;">DAA</h1>
<h2 style="font-family: 'Times New Roman', Times, serif; font-size: 40px;">Braga Weather from March 14th to April 4th</h2>
</center>

<font face="Times New Roman">

<h3 style="font-family: 'Times New Roman'">Imports</h3>

# Import necessary libraries
import pandas as pd              # Import pandas for data manipulation and analysis

<h3 style="font-family: 'Times New Roman'">Load Dataset Braga_2023-03-14_to_2023-04-04</h3>

df = pd.read_csv('Ficheiros/Ficheiros_Iniciais/Braga.csv')

<h3 style="font-family: 'Times New Roman'">Exploratory Data Analysis</h3>

# Get the column names as a list
column_names = df.columns.to_list()

# Print the column names one by one with a number
print('Columns:')
for i, column in enumerate(column_names, 1):
    print(f"{i} -> '{column}'")

# Display information about the DataFrame
df.info()

# Generate and print descriptive statistics for each column in the specified 'columns' list
for column in column_names:
    print(df[column].describe())
    print('\n')

# Print unique values for each column in the 'columns' list
for column in column_names:
    print(f'{column}: {df[column].unique()}')

<h3 style="font-family: 'Times New Roman'">Data Transformation</h3>

df.rename(columns={
    'dt_iso': 'date',
}, inplace=True)

df['date'] = df['date'].str.replace(' \+\d+ UTC', '', regex=True)
df['date'] = df['date'].str.replace(':[0-9][0-9]$', '', regex=True)
df['date'] = df['date'].str.replace('T', ' ', regex=True)

# Mais a frente nao tirar wind_deg e dew_point e weather_main
df = df.drop(columns=['timezone', 
                      'lat', 
                      'lon', 
                      'visibility', 
                      'sea_level', 
                      'grnd_level', 
                      'wind_gust', 
                      'rain_3h', 
                      'snow_1h', 
                      'snow_3h', 
                      'city_name', 
                      'dt', 
                      'weather_icon', 
                      'weather_id', 
                      'wind_deg',
                      'dew_point',
                      'weather_main'
                      ])

<h3 style="font-family: 'Times New Roman'">Save DataSet Braga</h3>

df.to_csv('Ficheiros/Ficheiros_Finais/Braga.csv', index = False)
