==> Source/DataSetCompeticao/2-Data_Combination.ipynb <==
<center>
<h1 style="font-family: 'Times New Roman', Times, serif; font-size: 60px;">DAA</h1>
<h2 style="font-family: 'Times New Roman', Times, serif; font-size: 40px;">Data Combination</h2>
</center>

<font face="Times New Roman">

<h3 style="font-family: 'Times New Roman'">Imports</h3>

# Import necessary libraries
import pandas as pd              # Import pandas for data manipulation and analysis

<h3 style="font-family: 'Times New Roman'">Load All Datasets</h3>

meteo21 = pd.read_csv('Ficheiros/Ficheiros_Iniciais/meteo_202109-202112.csv')

meteo22 = pd.read_csv('Ficheiros/Ficheiros_Iniciais/meteo_202201-202212.csv')

meteoteste = pd.read_csv('Ficheiros/Ficheiros_Iniciais/meteo_202301-202304.csv')

energy21 = pd.read_csv('Ficheiros/Ficheiros_Iniciais/energia_202109-202112.csv', na_filter = False)

energy22 = pd.read_csv('Ficheiros/Ficheiros_Iniciais/energia_202201_202212.csv', na_filter = False)

energyteste = pd.read_csv('Ficheiros/Ficheiros_Iniciais/energia_202301-202304.csv', na_filter = False)

braga = pd.read_csv("Ficheiros/Ficheiros_Finais/Braga.csv")

<h3 style="font-family: 'Times New Roman'">Exploratory Data Analysis</h3>

# Get the column names as a list
column_names_meteo21 = meteo21.columns.to_list()

# Print the column names one by one with a number
print('Columns:')
for i, column in enumerate(column_names_meteo21, 1):
    print(f"{i} -> '{column}'")

# Get the column names as a list
column_names_meteo22 = meteo22.columns.to_list()

# Print the column names one by one with a number
print('Columns:')
for i, column in enumerate(column_names_meteo22, 1):
    print(f"{i} -> '{column}'")

# Get the column names as a list
column_names_meteoteste = meteoteste.columns.to_list()

# Print the column names one by one with a number
print('Columns:')
for i, column in enumerate(column_names_meteoteste, 1):
    print(f"{i} -> '{column}'")

# Get the column names as a list
column_names_braga = braga.columns.to_list()

# Print the column names one by one with a number
print('Columns:')
for i, column in enumerate(column_names_braga, 1):
    print(f"{i} -> '{column}'")

# Get the column names as a list
column_names_energy21 = energy21.columns.to_list()

# Print the column names one by one with a number
print('Columns:')
for i, column in enumerate(column_names_energy21, 1):
    print(f"{i} -> '{column}'")

# Get the column names as a list
column_names_energy22 = energy22.columns.to_list()

# Print the column names one by one with a number
print('Columns:')
for i, column in enumerate(column_names_energy22, 1):
    print(f"{i} -> '{column}'")

# Get the column names as a list
column_names_energyteste = energyteste.columns.to_list()

# Print the column names one by one with a number
print('Columns:')
for i, column in enumerate(column_names_energyteste, 1):
    print(f"{i} -> '{column}'")

meteo21.info()

meteo22.info()

meteoteste.info()

braga.info()

energy21.info()

energy22.info()

energyteste.info()

<h3 style="font-family: 'Times New Roman'">Data Transformation</h3>

meteo21.rename(columns={
    'dt': 'timestamp',
    'dt_iso': 'date',
    'city_name': 'city',
}, inplace=True)

meteo22.rename(columns={
    'dt': 'timestamp',
    'dt_iso': 'date',
    'city_name': 'city',
}, inplace=True)

meteoteste.rename(columns={
    'dt': 'timestamp',
    'dt_iso': 'date',
    'city_name': 'city',
}, inplace=True)

meteo21['date'] = meteo21['date'].str.replace(' \+\d+ UTC', '', regex=True)
meteo21['date'] = meteo21['date'].str.replace(':[0-9][0-9]$', '', regex=True)

meteo22['date'] = meteo22['date'].str.replace(' \+\d+ UTC', '', regex=True)
meteo22['date'] = meteo22['date'].str.replace(':[0-9][0-9]$', '', regex=True)

meteoteste['date'] = meteoteste['date'].str.replace(' \+\d+ UTC', '', regex=True)
meteoteste['date'] = meteoteste['date'].str.replace(':[0-9][0-9]$', '', regex=True)

energy21.rename(columns={
    'Data': 'date',
    'Hora': 'hour',
    'Normal (kWh)': 'normal_consume',
    'Horario Economico (kWh)': 'consumption_in_period',
    'Autoconsumo (kWh)': 'autoconsume',
    'Injecao na rede (kWh)': 'injection',
}, inplace=True)

energy22.rename(columns={
    'Data': 'date',
    'Hora': 'hour',
    'Normal (kWh)': 'normal_consume',
    'Horario Economico (kWh)': 'consumption_in_period',
    'Autoconsumo (kWh)': 'autoconsume',
    'Injecao na rede (kWh)': 'injection',
}, inplace=True)

energyteste.rename(columns={
    'Data': 'date',
    'Hora': 'hour',
    'Normal (kWh)': 'normal_consume',
    'Horario Economico (kWh)': 'consumption_in_period',
    'Autoconsumo (kWh)': 'autoconsume',
    'Injecao na rede (kWh)': 'injection',
}, inplace=True)

energy21['hour'] = energy21['hour'].astype(str).str.zfill(2)
energy21['date'] = energy21['date'] + ' ' + energy21['hour'] + ':00'
energy21 = energy21.drop(columns=['hour'])

energy22['hour'] = energy22['hour'].astype(str).str.zfill(2)
energy22['date'] = energy22['date'] + ' ' + energy22['hour'] + ':00'
energy22 = energy22.drop(columns=['hour'])

energyteste['hour'] = energyteste['hour'].astype(str).str.zfill(2)
energyteste['date'] = energyteste['date'] + ' ' + energyteste['hour'] + ':00'
energyteste = energyteste.drop(columns=['hour'])

meteo21 = meteo21.drop(columns=['city',
                                   'timestamp',
                                   'sea_level',
                                   'grnd_level'])

meteo22 = meteo22.drop(columns=['city',
                                      'timestamp',
                                      'sea_level',
                                      'grnd_level'])

meteoteste = meteoteste.drop(columns=['city',
                                      'timestamp',
                                      'sea_level',
                                      'grnd_level'])

<h3 style="font-family: 'Times New Roman'">Exploratory Data Analysis</h3>

meteo21.columns

meteo22.columns

meteoteste.columns

braga.columns

energy21.columns

energy22.columns

energyteste.columns

<h3 style="font-family: 'Times New Roman'">Save all DataSets</h3>

resulting_df_21 = energy21.merge(meteo21, on='date', how='left')

resulting_df_22 = energy22.merge(meteo22, on='date', how='inner')

df_new = pd.read_csv("Ficheiros/Ficheiros_Iniciais/DewPoint.csv")

df_new_teste = pd.read_csv("Ficheiros/Ficheiros_Iniciais/DewPoint_teste.csv")

final = pd.concat([resulting_df_21, resulting_df_22], ignore_index=True)

resulting_dew_point = final.merge(df_new, on='date', how='inner')
resulting_dew_point.to_csv("Ficheiros/Ficheiros_Finais/Final.csv", index = False)

meteoteste = pd.concat([meteoteste, braga])

resulting_df_23 = energyteste.merge(meteoteste, on='date', how='inner')

resulting_dew_point_teste = resulting_df_23.merge(df_new_teste, on='date', how='inner')
resulting_dew_point_teste.to_csv("Ficheiros/Ficheiros_Finais/Final_Teste.csv", index = False)
