==> Source/DataSetCompeticao/4-Pre_Processing.ipynb <==
<center>
<h1 style="font-family: 'Times New Roman', Times, serif; font-size: 60px;">DAA</h1>
<h2 style="font-family: 'Times New Roman', Times, serif; font-size: 40px;">Pre Processing</h2>
</center>

<font face="Times New Roman">

<h3 style="font-family: 'Times New Roman'">Imports</h3>

# Import necessary libraries
import pandas as pd              # Import pandas for data manipulation and analysis
import numpy as np               # Import numpy for numerical computations

<h3 style="font-family: 'Times New Roman'">Load All Datasets</h3>

<!-- <h3 style="font-family: 'Times New Roman'">Load All Datasets</h3> -->

final = pd.read_csv("Ficheiros/Ficheiros_Pos_DataAnalysis_Before_PP/Final.csv", na_filter=False)
final_teste = pd.read_csv("Ficheiros/Ficheiros_Pos_DataAnalysis_Before_PP/Final_Teste.csv")

# pd.set_option('display.max_columns', 500)
# pd.set_option('display.max_rows', 500)

<h3 style="font-family: 'Times New Roman'">Pre Processing</h3>

# Filter rows where 'consumption_in_period' is between 6 and 8 (inclusive)
filtered_rows = final[(final['consumption_in_period'] >= 6) & (final['consumption_in_period'] <= 8)]

# Display the resulting DataFrame
print(filtered_rows)

# Drop rows where 'consumption_in_period' is between 6 and 8 (inclusive)
final.drop(final[(final['consumption_in_period'] >= 6) & (final['consumption_in_period'] <= 8)].index, inplace=True)

final['consumo_total'] = (final['autoconsume'] + final['normal_consume'] + final['consumption_in_period']).round(2)
final_teste['consumo_total'] = (final_teste['autoconsume'] + final_teste['normal_consume'] + final_teste['consumption_in_period']).round(2)

mask = final['consumo_total'] != 0
final['taxa_autoconsumo'] = np.where(mask, (final['autoconsume'] / final['consumo_total']).round(2), 0)

mask2 = final_teste['consumo_total'] != 0
final_teste['taxa_autoconsumo'] = np.where(mask2, (final_teste['autoconsume'] / final_teste['consumo_total']).round(2), 0)

# Extracting the hour part as an integer
final['hour'] = final['hour_str'].str.extract('(\d+)', expand=False).astype(int)

# Creating the new column based on the conditions
final['hora_partida'] = (final['hour'] >= 7) & (final['hour'] <= 18)

# Converting the boolean results to 1 and 0
final['hora_partida'] = final['hora_partida'].astype(int)

# Extracting the hour part as an integer
final_teste['hour'] = final_teste['hour_str'].str.extract('(\d+)', expand=False).astype(int)

# Creating the new column based on the conditions
final_teste['hora_partida'] = (final_teste['hour'] >= 7) & (final_teste['hour'] <= 18)

# Converting the boolean results to 1 and 0
final_teste['hora_partida'] = final_teste['hora_partida'].astype(int)

final['injection_tentativa'] = (final['autoconsume'] - (final['normal_consume'] + final['consumption_in_period'])).round(2)
final_teste['injection_tentativa'] = (final_teste['autoconsume'] - (final_teste['normal_consume'] + final_teste['consumption_in_period'])).round(2)

final.columns

final["hour_str"]

final['date'] = pd.to_datetime(final['date'])
final_teste['date'] = pd.to_datetime(final_teste['date'])

final.dtypes

import ephem

# Criar colunas auxiliares
final['year'] = final['date'].dt.year
final['month'] = final['date'].dt.month
final['day'] = final['date'].dt.day
final['hour'] = final['date'].dt.hour

# Converter 'hour' para string formatada com dois dígitos
final['hour_str'] = final['hour'].apply(lambda x: f'{x:02d}')

# Converter 'hour_str' para inteiro
final['hour_str'] = final['hour_str'].astype(int)


# Função para verificar se está ensolarado
def estava_sol(row):
    observador = ephem.Observer()
    observador.lat = '41.545449'  # Substitua pela latitude desejada
    observador.lon = '-8.426507'  # Substitua pela longitude desejada

    data_hora = ephem.Date(f'{row["year"]:04d}/{row["month"]:02d}/{row["day"]:02d} {row["hour_str"]:02d}:00')
    observador.date = data_hora

    sol = ephem.Sun()
    sol.compute(observador)

    # Retorna 1 se estiver ensolarado, 0 caso contrário
    return int(sol.alt > 0)

# Aplicar a função ao DataFrame
final['is_sunny'] = final.apply(estava_sol, axis=1)

import ephem

# Criar colunas auxiliares
final_teste['year'] = final_teste['date'].dt.year
final_teste['month'] = final_teste['date'].dt.month
final_teste['day'] = final_teste['date'].dt.day
final_teste['hour'] = final_teste['date'].dt.hour

# Converter 'hour' para string formatada com dois dígitos
final_teste['hour_str'] = final_teste['hour'].apply(lambda x: f'{x:02d}')

# Converter 'hour_str' para inteiro
final_teste['hour_str'] = final_teste['hour_str'].astype(int)


# Função para verificar se está ensolarado
def estava_sol(row):
    observador = ephem.Observer()
    observador.lat = '41.545449'  # Substitua pela latitude desejada
    observador.lon = '-8.426507'  # Substitua pela longitude desejada

    data_hora = ephem.Date(f'{row["year"]:04d}/{row["month"]:02d}/{row["day"]:02d} {row["hour_str"]:02d}:00')
    observador.date = data_hora

    sol = ephem.Sun()
    sol.compute(observador)

    # Retorna 1 se estiver ensolarado, 0 caso contrário
    return int(sol.alt > 0)


# Aplicar a função ao DataFrame
final_teste['is_sunny'] = final_teste.apply(estava_sol, axis=1)

<h3 style="font-family: 'Times New Roman'">Save DataSet</h3>

final.to_csv("Ficheiros/Ficheiros_Pos_Pre_Processing/Final.csv", index = False)
final_teste.to_csv("Ficheiros/Ficheiros_Pos_Pre_Processing/Final_Teste.csv", index = False)
